Move getPostTypeStaticPath to functions
