+++ functions/wordpress/postTypes/getPostTypeStaticPaths.js
